WebCryptoAPI/RSA-OAEP.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>RSA-OAEP</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="util.js"></script>
<script src="test-vectors.js"></script>
<div id=log></div>
<script>
"use strict";

promise_test(function() {
  var alg = {name: "RSA-OAEP", hash: "SHA-1"};

  return Promise.all([
    crypto.subtle.importKey("pkcs8", tv.rsaoaep.pkcs8, alg, false, ['decrypt']),
    crypto.subtle.importKey("spki", tv.rsaoaep.spki, alg, false, ['encrypt'])
      .then(function(pubKey) { return crypto.subtle.encrypt(alg, pubKey, tv.rsaoaep.data); })
  ]).then(function(values) {
    var privKey = values[0];
    var encrypted = values[1];
    return crypto.subtle.decrypt(alg, privKey, encrypted);
  }).then(function(x) {
    assert_true(util.memcmp(x, tv.rsaoaep.data));
  });
}, "RSA-OAEP encrypt/decrypt round-trip");

/*
// -----------------------------------------------------------------------------
TestArray.addTest(
  "RSA-OAEP key generation and encrypt/decrypt round-trip (SHA-256)",
  function () {
    var that = this;
    var alg = {
      name: "RSA-OAEP",
      hash: "SHA-256",
      modulusLength: 2048,
      publicExponent: new Uint8Array([0x01, 0x00, 0x01])
    };

    var privKey, pubKey, data = crypto.getRandomValues(new Uint8Array(128));
    function setKey(x) { pubKey = x.publicKey; privKey = x.privateKey; }
    function doEncrypt() {
      return crypto.subtle.encrypt(alg, pubKey, data);
    }
    function doDecrypt(x) {
      return crypto.subtle.decrypt(alg, privKey, x);
    }

    crypto.subtle.generateKey(alg, false, ['encrypt', 'decrypt'])
      .then(setKey, error(that))
      .then(doEncrypt, error(that))
      .then(doDecrypt, error(that))
      .then(
        memcmp_complete(that, data),
        error(that)
      );
  }
);

// -----------------------------------------------------------------------------
TestArray.addTest(
  "RSA-OAEP decryption known answer",
  function () {
    var that = this;
    var alg = {name: "RSA-OAEP", hash: "SHA-1"};

    function doDecrypt(x) {
      return crypto.subtle.decrypt(alg, x, tv.rsaoaep.result);
    }
    function fail() { error(that); }

    crypto.subtle.importKey("pkcs8", tv.rsaoaep.pkcs8, alg, false, ['decrypt'])
      .then( doDecrypt, fail )
      .then( memcmp_complete(that, tv.rsaoaep.data), fail );
  }
);

// -----------------------------------------------------------------------------
TestArray.addTest(
  "RSA-OAEP input data length checks (2048-bit key)",
  function () {
    var that = this;
    var privKey, pubKey;
    var alg = {
      name: "RSA-OAEP",
      hash: "SHA-1",
      modulusLength: 2048,
      publicExponent: new Uint8Array([0x01, 0x00, 0x01])
    };

    var privKey, pubKey;
    function setKey(x) { pubKey = x.publicKey; privKey = x.privateKey; }
    function doEncrypt(n) {
      console.log("entered encrypt("+ n +")");
      return function () {
        return crypto.subtle.encrypt(alg, pubKey, new Uint8Array(n));
      }
    }

    crypto.subtle.generateKey(alg, false, ['encrypt', 'decrypt'])
      .then(setKey, error(that))
      .then(doEncrypt(214), error(that))
      .then(doEncrypt(215), error(that))
      .then(error(that), complete(that));
  }
);

// -----------------------------------------------------------------------------
TestArray.addTest(
  "RSA-OAEP key import with invalid hash",
  function () {
    var that = this;
    var alg = {name: "RSA-OAEP", hash: "SHA-123"};

    crypto.subtle.importKey("pkcs8", tv.rsaoaep.pkcs8, alg, false, ['decrypt'])
      .then(error(that), complete(that));
  }
);

// -----------------------------------------------------------------------------
TestArray.addTest(
  "Test that RSA-OAEP encrypt/decrypt accepts strings as AlgorithmIdentifiers",
  function () {
    var that = this;
    var alg = {
      name: "RSA-OAEP",
      hash: "SHA-256",
      modulusLength: 2048,
      publicExponent: new Uint8Array([0x01, 0x00, 0x01])
    };

    var privKey, pubKey, data = crypto.getRandomValues(new Uint8Array(128));
    function setKey(x) { pubKey = x.publicKey; privKey = x.privateKey; }
    function doEncrypt() {
      return crypto.subtle.encrypt("RSA-OAEP", pubKey, data);
    }
    function doDecrypt(x) {
      return crypto.subtle.decrypt("RSA-OAEP", privKey, x);
    }

    crypto.subtle.generateKey(alg, false, ["encrypt", "decrypt"])
      .then(setKey)
      .then(doEncrypt)
      .then(doDecrypt)
      .then(memcmp_complete(that, data), error(that));
  }
);
*/
</script>
